requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
Werkzeug==3.0.1
marshmallow==3.20.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2
requests==2.31.0
geopy==2.4.1
APScheduler==3.10.4
Pillow==10.1.0
python-dateutil==2.8.2
bcrypt==4.1.2
gunicorn==21.2.0
